host/requirements.txt
pyserial
docker